Refactor: VideoChat 리패터링 작업 [1]
