+++ src/components/VideoChat/VideoLayout.tsx
@@ -1,0 +1,45 @@
+import React from "react";
+
+const VideoLayout = ({
+  localVideoRef,
+  remoteVideoRef,
+  isLocalScreenSharing,
+  isRemoteScreenSharing,
+}) => {
+  return (
+    <div className="VideoChattingWrapper-VideoWrapper">
+      <video
+        className={`VideoBox ${isLocalScreenSharing ? "Sharing" : ""}`}
+        ref={localVideoRef}
+        autoPlay
+        playsInline
+        muted
+      >
+        <svg
+          className="VideoBox-expendIcon"
+          xmlns="http://www.w3.org/2000/svg"
+          viewBox="0 0 448 512"
+        >
+          <path d="M32 32C14.3 32 0 46.3 0 64l0 96c0 17.7 14.3 32 32 32s32-14.3 32-32l0-64 64 0c17.7 0 32-14.3 32-32s-14.3-32-32-32L32 32zM64 352c0-17.7-14.3-32-32-32s-32 14.3-32 32l0 96c0 17.7 14.3 32 32 32l96 0c17.7 0 32-14.3 32-32s-14.3-32-32-32l-64 0 0-64zM320 32c-17.7 0-32 14.3-32 32s14.3 32 32 32l64 0 0 64c0 17.7 14.3 32 32 32s32-14.3 32-32l0-96c0-17.7-14.3-32-32-32l-96 0zM448 352c0-17.7-14.3-32-32-32s-32 14.3-32 32l0 64-64 0c-17.7 0-32 14.3-32 32s14.3 32 32 32l96 0c17.7 0 32-14.3 32-32l0-96z" />
+        </svg>
+      </video>
+      <video
+        className={`VideoBox ${isRemoteScreenSharing ? "Sharing" : ""}`}
+        ref={remoteVideoRef}
+        autoPlay
+        playsInline
+        muted
+      >
+        <svg
+          className="VideoBox-expendIcon"
+          xmlns="http://www.w3.org/2000/svg"
+          viewBox="0 0 448 512"
+        >
+          <path d="M32 32C14.3 32 0 46.3 0 64l0 96c0 17.7 14.3 32 32 32s32-14.3 32-32l0-64 64 0c17.7 0 32-14.3 32-32s-14.3-32-32-32L32 32zM64 352c0-17.7-14.3-32-32-32s-32 14.3-32 32l0 96c0 17.7 14.3 32 32 32l96 0c17.7 0 32-14.3 32-32s-14.3-32-32-32l-64 0 0-64zM320 32c-17.7 0-32 14.3-32 32s14.3 32 32 32l64 0 0 64c0 17.7 14.3 32 32 32s32-14.3 32-32l0-96c0-17.7-14.3-32-32-32l-96 0zM448 352c0-17.7-14.3-32-32-32s-32 14.3-32 32l0 64-64 0c-17.7 0-32 14.3-32 32s14.3 32 32 32l96 0c17.7 0 32-14.3 32-32l0-96z" />
+        </svg>
+      </video>
+    </div>
+  );
+};
+
+export default VideoLayout;
